Add repeat event exceptions

+++ src/views/event/eventView/EventView.tsx
@@ -12,12 +12,14 @@ import {
 import { createToast, formatEventDate } from '../../../utils/common';
 
 import { Context } from 'context/store';
-import { EVENT_TYPE } from 'bloben-interface/enums';
+import { EVENT_TYPE, REPEATED_EVENT_CHANGE_TYPE } from 'bloben-interface/enums';
 import { EvaIcons } from 'components/eva-icons';
 import { Stack, Text, useToast } from '@chakra-ui/react';
 import { TOAST_STATUS } from '../../../types/enums';
 import { WebcalCalendar } from '../../../redux/reducers/webcalCalendars';
 import { calendarByEvent } from '../../../utils/tsdavHelper';
+import { debug } from '../../../utils/debug';
+import { v4 } from 'uuid';
 import CalDavEventsApi from '../../../api/CalDavEventsApi';
 import EventDetailAttendee from '../../../components/eventDetail/eventDetailAttendee/EventDetailAttendee';
 import EventDetailCalendar from '../../../components/eventDetail/eventDetailCalendar/EventDetailCalendar';
@@ -26,7 +28,19 @@ import EventDetailNotes from '../../../components/eventDetail/eventDetailNotes/E
 import EventDetailTitle from '../../../components/eventDetail/eventDetailTitle/EventDetailTitle';
 import FormIcon from '../../../components/formIcon/FormIcon';
 import HeaderModal from '../../../components/headerModal/HeaderModal';
+import ICalHelper from '../../../utils/ICalHelper';
 import Modal from 'components/modal/Modal';
+import RepeatEventModal, {
+  REPEAT_MODAL_TYPE,
+} from '../../../components/repeatEventModal/RepeatEventModal';
+
+export const checkIfHasRepeatPreAction = (event: any) => {
+  return (
+    event?.rRule?.length > 1 ||
+    event?.recurrenceID ||
+    event?.recurrenceID?.value
+  );
+};
 
 interface EventDatesProps {
   event: CalDavEvent;
@@ -74,6 +88,7 @@ const EventView = (props: EventViewProps) => {
   // const [isOrganizer, setIsOrganizer] = useState(false);
   const [event, setEvent] = useState(null as any);
   const [calendar, setCalendar] = useState(null as any);
+  const [deleteModalOpen, openDeleteModal] = useState(false);
 
   const calendars: CalDavCalendar[] = useSelector(
     (state: ReduxState) => state.calDavCalendars
@@ -104,6 +119,11 @@ const EventView = (props: EventViewProps) => {
   };
 
   const deleteEvent = async () => {
+    if (checkIfHasRepeatPreAction(event) && !deleteModalOpen) {
+      openDeleteModal(true);
+      return;
+    }
+
     try {
       const response = await CalDavEventsApi.deleteEvent({
         calendarID: calendar.id,
@@ -112,6 +132,91 @@ const EventView = (props: EventViewProps) => {
         id: event.id,
       });
       if (response.status === 200 || response.status === 204) {
+        setContext('syncSequence', store.syncSequence + 1);
+
+        toast(createToast('Event deleted'));
+
+        handleClose();
+      }
+    } catch (e: any) {
+      toast(createToast(e.response?.data?.message, TOAST_STATUS.ERROR));
+    }
+  };
+
+  const handleDeleteRepeated = async (value: REPEATED_EVENT_CHANGE_TYPE) => {
+    try {
+      // use issued id or create for new event
+      const newEventExternalID: string = event?.externalID || v4();
+
+      let iCalString: string | null;
+
+      let response;
+
+      if (value === REPEATED_EVENT_CHANGE_TYPE.SINGLE) {
+        if (event.recurrenceID || event.recurrenceID?.value) {
+          response = await CalDavEventsApi.deleteRepeatedEvent({
+            calendarID: calendar.id,
+            url: event.url,
+            etag: event.etag,
+            id: event.id,
+            type: REPEATED_EVENT_CHANGE_TYPE.SINGLE_RECURRENCE_ID,
+            recurrenceID: event.recurrenceID,
+          });
+          if (response.status === 200 || response.status === 204) {
+            setContext('syncSequence', store.syncSequence + 1);
+
+            toast(createToast('Event deleted'));
+
+            handleClose();
+          }
+
+          return;
+        } else {
+          const originalEvent = await CalDavEventsApi.getEvent(
+            calendar.id,
+            event.url
+          );
+          iCalString = new ICalHelper({
+            ...event,
+            startAt: originalEvent.data.startAt,
+            endAt: originalEvent.data.endAt,
+            timezoneStart: originalEvent.data.timezoneStart,
+            externalID: newEventExternalID,
+            rRule: originalEvent.data.rRule,
+            recurrenceID: undefined,
+            exdates: [
+              ...event.exdates,
+              { value: event.startAt, timezone: event.timezoneStartAt },
+            ],
+          }).parseTo();
+
+          debug(iCalString);
+
+          response = await CalDavEventsApi.deleteRepeatedEvent({
+            calendarID: calendar.id,
+            url: event.url,
+            etag: event.etag,
+            id: event.id,
+            type: value,
+            // @ts-ignore
+            iCalString: iCalString ? iCalString : null,
+            exDates: [
+              ...event.exdates,
+              { value: event.startAt, timezone: event.timezoneStartAt },
+            ],
+          });
+        }
+      } else if (value === REPEATED_EVENT_CHANGE_TYPE.ALL) {
+        response = await CalDavEventsApi.deleteRepeatedEvent({
+          calendarID: calendar.id,
+          url: event.url,
+          etag: event.etag,
+          id: event.id,
+          type: value,
+        });
+      }
+
+      if (response?.status === 200 || response?.status === 204) {
         setContext('syncSequence', store.syncSequence + 1);
 
         toast(createToast('Event deleted'));
@@ -132,7 +237,14 @@ const EventView = (props: EventViewProps) => {
     getCalendar();
   }, [JSON.stringify(props.data)]);
 
-  return event && event.id ? (
+  return checkIfHasRepeatPreAction(event) && deleteModalOpen ? (
+    <RepeatEventModal
+      type={REPEAT_MODAL_TYPE.DELETE}
+      handleClose={handleClose}
+      title={''}
+      handleClick={handleDeleteRepeated}
+    />
+  ) : event && event.id ? (
     <Modal e={currentE} handleClose={handleClose} maxHeight={'42%'}>
       <>
         {event.type === EVENT_TYPE.CALDAV ? (
